simplify & make less blinding
I noticed the scattering tended to be way too bright during the day, so I modified the code a bit more. It is a little less dramatic at sunset and sunrises now, but is way less aggressive, which is nice. Also cleaner. And anyone who wants more dramatic scattering can turn up the slider anyway